% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoreject.by.absolute.threshold.R
\name{autoreject.by.absolute.threshold}
\alias{autoreject.by.absolute.threshold}
\title{Identify outlier observations based on absolute threshold}
\usage{
autoreject.by.absolute.threshold(x, threshold, dir)
}
\arguments{
\item{x}{A matrix or data frame containing the time series data. Each column represents
a different repeat, and each row represents a time point.}

\item{threshold}{A single numeric value of class units specifying the threshold for rejection.}

\item{dir}{A string value indicating whether to reject columns with measurements above ("above") or below ("below") the specified threshold, or with an absolute value above the threshold ("both").}
}
\value{
A logical vector indicating which observations to reject.
}
\description{
This function rejects entire columns in a matrix of class units if any value
in the column is above or below a specified absolute threshold.
}
\details{
The function iterates through each column of the matrix 'x' and marks
columns for rejection if any measurement in the column is above or below the specified threshold,
based on the 'dir' parameter.
}
\examples{
require(units)
data <- set_units(matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), ncol = 3), "m")
result <- autoreject.by.absolute.threshold(data, threshold = set_units(5, "m"), dir = "both")

}
