% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoreject.by.distance.R
\name{autoreject.by.distance}
\alias{autoreject.by.distance}
\title{Identify outlier observations  based on distances}
\usage{
autoreject.by.distance(x, threshold = 1)
}
\arguments{
\item{x}{A matrix or data frame containing the time series data. Each column represents
a different repeat, and each row represents a time point.}

\item{threshold}{The distance threshold for rejection (default is 1).}
}
\value{
A logical vector indicating which observations to reject.
}
\description{
This function calculates the Euclidean distance between observations in a matrix,
scales the data, and identifies outlier observations to reject based on a given threshold.
}
\details{
The function calculates the Euclidean distance between all pairs of
observations in the input matrix 'x'. It then scales the distance matrix and
identifies observations with mean distances greater than the specified threshold
for rejection.
}
\examples{
data <- matrix(data = c(1, 2, 3, 4, 5, 6, 7, 8, 9), ncol = 3)
rejected <- autoreject.by.distance(data, threshold = 1)

}
