% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoreject.by.PCA.R
\name{autoreject.by.PCA}
\alias{autoreject.by.PCA}
\title{Identify outlier observations based on PCA loadings}
\usage{
autoreject.by.PCA(x, pct.keep = 0.75)
}
\arguments{
\item{x}{A matrix or data frame with the observations.}

\item{pct.keep}{The proportion of observations to keep based on their PCA loadings (default is 0.75).}
}
\value{
A logical vector indicating which observations to reject.
}
\description{
This function performs Principal Component Analysis (PCA) on a matrix or data frame,
scales the data, and identifies outlier observations to reject based on their contributions
to the principal components. The observations with the highest contribution to the principal
components are considered outliers.
}
\details{
The function performs PCA on the input matrix 'x', scales the data, and calculates
the loadings of the observations. It then identifies the top contributing observations to the
principal components based on a specified proportion to keep. Observations with lower contributions
are marked for rejection.
}
\examples{
# data <- matrix(data = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12), ncol = 3)
# rejected <- autoreject.by.PCA(data, pct.keep = 0.75)

}
\keyword{internal}
