% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoreject.by.signalfree.R
\name{autoreject.by.signalfree}
\alias{autoreject.by.signalfree}
\title{Detect and Outlier Traces in Time Series Data with repeated measurements}
\usage{
autoreject.by.signalfree(x, rejection.cutoff = 1)
}
\arguments{
\item{x}{A matrix or data frame containing the time series data. Each column represents
a different repeat, and each row represents a time point.}

\item{rejection.cutoff}{A numeric value representing the threshold for outlier rejection.}
}
\value{
A logical vector indicating whether each time series is rejected (TRUE) or not (FALSE).
}
\description{
This function detects and rejects outlier traces in time series data based on signal-free areas in the trace (e.g. before s signal-causing even occurred or when the response to it has ended).
}
\details{
The function applies a series of data processing steps, including data scaling,
calculation of areas of maximum normalized spread, smoothening, and scaling, to identify
areas of high variance. Time series with high variance areas exceeding the rejection
cutoff are considered as outliers and are marked for rejection.
}
\examples{
# Example usage
data_matrix <- matrix(rnorm(30000), ncol = 30)
rejected <- autoreject.by.signalfree(data_matrix, rejection.cutoff = 1)

}
