% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_unit.R
\name{convert_to_unit}
\alias{convert_to_unit}
\title{Convert a variable to a specified unit}
\usage{
convert_to_unit(variable, target_unit)
}
\arguments{
\item{variable}{A variable of class "units" to be converted.}

\item{target_unit}{A string specifying the target unit to which the conversion should be attempted.}
}
\value{
Returns the variable converted to the specified unit.
}
\description{
This function attempts to convert a variable of class "units" to a specified unit.
It checks if the input variable has the correct class and handles errors during conversion.
}
\examples{
require(units)
my_variable <- set_units(1, "minutes")
tryCatch({
  converted_variable <- convert_to_unit(my_variable, "seconds")
  print(converted_variable)
}, error = function(e) {
  print(e$message)
})
}
