% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfourier.R
\name{fastfourier}
\alias{fastfourier}
\title{Compute the Fast Fourier Transform (FFT) of a signal.}
\usage{
fastfourier(x, samp.freq, ...)
}
\arguments{
\item{x}{Numeric vector or matrix of signal data.}

\item{samp.freq}{Numeric scalar, the sampling frequency in Hertz.}

\item{...}{Additional arguments to pass to \code{\link[stats]{fft}}.}
}
\value{
A data frame with two columns: 'fur' representing the Fourier
coefficients and 'freq' representing the corresponding frequencies.
}
\description{
This function calculates the FFT of a signal and returns the frequency domain
representation.
}
\examples{
fft_result <- fastfourier(signal, samp.freq = 1000)

}
