% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.bandpass.R
\name{filter.bandpass}
\alias{filter.bandpass}
\title{Filter a signal with a bandpass filter.}
\usage{
filter.bandpass(x, samp.freq, low, high)
}
\arguments{
\item{x}{Numeric vector or matrix of signal data.}

\item{samp.freq}{Numeric scalar, the sampling frequency in Hertz.}

\item{low}{Numeric scalar, the lower cutoff frequency in Hertz.}

\item{high}{Numeric scalar, the upper cutoff frequency in Hertz.}
}
\value{
Filtered signal with the same dimensions as input 'x'.
}
\description{
This function filters a signal using a fourth order Butterworth bandpass filter with
specified low and high frequency cutoffs. This function uses the \link[signal:butter]{signal::butter} function from the \link[signal:signal]{signal::signal} package.
}
\examples{
signal<-sin((1:1000)/30)*sin((1:1000)/60)
plot(signal)
filtered_signal <- filter.bandpass(signal, samp.freq = 1000, low = 5, high = 100)
plot(filtered_signal)
}
\seealso{
\link[signal:butter]{signal::butter}
}
