% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.detrend.R
\name{filter.detrend}
\alias{filter.detrend}
\title{Apply Detrending Filter}
\usage{
\method{filter}{detrend}(x, ...)
}
\arguments{
\item{x}{A numeric vector representing the time series data.}

\item{...}{Additional arguments to be passed to \code{\link[pracma:detrend]{pracma::detrend}} function.}
}
\value{
A numeric vector representing the detrended time series.
}
\description{
This function is a wrapper for the \code{\link[pracma:detrend]{pracma::detrend}} function in the pracma package.
It applies a detrending filter to a given time series.
}
\examples{
x <- c(1, 2, 3, 4, 3, 2, 1)
detrended <- filter.detrend(x)

}
\seealso{
\code{\link[pracma:detrend]{detrend::detrend}}

Other filter functions: 
\code{\link{filter.lin.detrend}()}
}
\concept{filter functions}
