% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.lin.detrend.R
\name{filter.lin.detrend}
\alias{filter.lin.detrend}
\title{Apply Linear Detrending Filter}
\usage{
\method{filter}{lin.detrend}(x)
}
\arguments{
\item{x}{A numeric vector representing the time series data.}
}
\value{
A numeric vector representing the detrended time series.
}
\description{
This function substracts the linear trend from the data.
}
\examples{
x <- c(1, 2, 3, 4, 3, 2, 1)
detrended <- filter.detrend(x)

}
\seealso{
\code{\link[pracma]{detrend}}

Other filter functions: 
\code{\link{filter.detrend}()}
}
\concept{filter functions}
