% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.to.w.R
\name{freq.to.w}
\alias{freq.to.w}
\title{Convert Frequency to 'Angular' Frequency (input to signalbutterworth filter)}
\usage{
freq.to.w(x, time.trace = NULL, samp.freq = NULL)
}
\arguments{
\item{x}{Numeric vector. Frequency values to be converted. If no units (\link[units:units]{units::units}) are provided, assumes Herz.}

\item{time.trace}{Numeric vector of at least length 2 representing a time trace to calculate the sample frequency from. Must be of class 'units' (\link[units:units]{units::units}), with a unit convertible to 's'.}

\item{samp.freq}{Numeric. Sample frequency in Hz, or, if of class 'units' convertibel into 'Hz'.}
}
\value{
Numeric vector of the converted angular frequencies.
}
\description{
This function converts a frequency to 'angular' frequency \code{W}, as required by \link[signal:butter]{signal::butter}. Note this is not truly an angular frequency but in fact 'x / (samp.freq / 2)'.
}
\details{
This function checks if the input units are correct and converts the given frequency values to angular frequency using the provided time values or sample frequency. The function understands units (\link{units::units}), so values can be provided with units.

If both \code{time.trace} and \code{samp.freq} are provided, an error is thrown.

If only \code{time.trace} is provided, the function calculates the sample frequency from it. If only \code{samp.freq} is provided, it is directly used.
}
\examples{
freq.to.w(x = 10, samp.freq = 100) # Returns 0.314159265358979
require(units)
freq.to.w(x = 10, time.trace = as_units(c(0, 1, 2),"s")) # Returns 3.14159265358979

}
