% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastfourier.R
\name{fastfourier}
\alias{fastfourier}
\title{Compute the Discrete Fourier Transform (DFT) of a signal.}
\usage{
fastfourier(x, samp.freq, ...)
}
\arguments{
\item{x}{Numeric vector or matrix of signal data.}

\item{samp.freq}{Numeric scalar, the sampling frequency in Hertz.}

\item{...}{Additional arguments to pass to \code{\link[stats]{fft}}.}
}
\value{
A data frame with two columns: 'fur' representing the Fourier
coefficients and 'freq' representing the corresponding frequencies.
}
\description{
This function calculates the FFT of a signal and returns the frequency domain
representation.
}
\examples{
signal<-sin((1:1000)/33)*sin((1:1000)/20)
plot(signal)
fft_result <- fastfourier(signal, samp.freq = as_units(1000,"Hz"))
plot(fft_result$freq[1:20],abs(fft_result$fur[1:20]))
}
