% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.bandpass.R
\docType{Function}
\name{filter.bandpass}
\alias{filter.bandpass}
\title{Filter a signal with a bandpass filter.}
\usage{
\method{filter}{bandpass}(x, low, high)
}
\arguments{
\item{x}{Numeric vector or matrix of signal data.}

\item{low, high}{Numeric, Lower and upper filter bands, passed on as \code{W} to the \link[signal:butter]{signal::butter} function. Use \link{freq.to.w} to convert temporal frequencies into the required format.}
}
\value{
Filtered signal with the same dimensions as input 'x'.
}
\description{
This function filters a signal using a fourth order Butterworth band pass filter with
specified low and high frequency cutoffs. This function uses the \link[signal:butter]{signal::butter} function from the \link[signal:signal]{signal::signal} package.
}
\details{
\itemize{
\item For Mouse ERG recordings, low=1 and high=300 is recommended (https://open.fau.de/items/2548a085-59c9-45f7-9136-02a85b17ea51)
\item For Mouse VEP recordings, low=0.5 and high=100 is recommended (doi:10.1016/j.visres.2005.09.006)
}
}
\examples{
signal<-sin((1:1000)/30)*sin((1:1000)/60)
plot(signal)
filtered_signal <- filter.bandpass(signal, freq.to.w(x=6,samp.freq=1000), freq.to.w(x=20,samp.freq=1000))
plot(filtered_signal)
}
\seealso{
\link[signal:butter]{signal::butter}
}
