% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.detrend.R, R/filter.lin.detrend.R
\name{filter.detrend}
\alias{filter.detrend}
\alias{filter.lin.detrend}
\title{Apply Detrending Filter}
\usage{
filter.detrend(x, ...)

filter.lin.detrend(x)
}
\arguments{
\item{x}{A numeric vector representing the time series data.}

\item{...}{Additional arguments to be passed to \code{\link[pracma:detrend]{pracma::detrend}} function.}
}
\value{
A numeric vector representing the detrended time series.

A numeric vector representing the detrended time series.
}
\description{
This function is a wrapper for the \code{\link[pracma:detrend]{pracma::detrend}} function in the pracma package.
It applies a detrending filter to a given time series.

This function substracts the linear trend from the data.
}
\examples{
x <- c(1, 2, 3, 4, 3, 2, 1)
detrended <- filter.detrend(x)

x <- c(1, 2, 3, 4, 3, 2, 1)
detrended <- filter.detrend(x)

}
\seealso{
\code{\link[pracma]{detrend}}

\code{\link[pracma]{detrend}}
}
\concept{filter functions}
