% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSD.R
\name{PSD}
\alias{PSD}
\title{Calculate the Power Spectral Density (PSD) for Electrophysiological Data}
\usage{
PSD(X, bandpass = as_units(c(0.5, 300), "Hz"))
}
\arguments{
\item{X}{An object of class \code{EPhysData} data.}

\item{bandpass}{A numeric vector of two elements specifying the lower
and upper bounds of the frequency band of interest, expressed in Hz.}
}
\value{
An object of class \code{EPhysData} containing the PSD and corresponding frequency vector.
}
\description{
This method calculates the Power Spectral Density (PSD) of electrophysiological data
using a fast Fourier transform (FFT) and subsequent power calculations.
}
\details{
This method calculates the Power Spectral Density (PSD) of electrophysiological data
using a discrete Fourier transform (DFT) and subsequent power calculations.
The data is first bandpass filtered to retain only frequencies of interest.
Then, a DFT is applied to the filtered data to transform it into the frequency domain.
The PSD is computed from the squared magnitude of the DFT results.

The power spectral density is computed using the formula:
\deqn{PSD(f) = 2 \times \frac{1}{T \times N} \times |X(f)|^2}
where |X(f)|^2 is the squared magnitude of the DFT, \eqn{T} is the sampling period,
and \eqn{N} is the number of points in the DFT.
The factor of 2 accounts for the loss of power in the negative frequency components
which are not included in the single-sided PSD.
}
\examples{
require(EPhysData)
X<-makeExampleEPhysData()
ggEPhysData(X)
psd<-PSD(X)
ggEPhysData(psd)
}
